requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
black==25.9.0
claude-agent-sdk
click==8.3.0
flake8==7.3.0
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
pytokens==0.3.0
slack_bolt==1.26.0
slack_cli_hooks==0.2.0
slack_sdk==3.38.0
yarl==1.22.0
